separate the push test on different go version
